requirements.txt
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attr==0.3.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.20.0
bleach==5.0.1
boto==2.49.0
boto3==1.34.118
botocore==1.34.118
boxing==0.1.4
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
expiringdict==1.2.2
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.64.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
humansignal-drf-yasg==1.21.9
idna==3.7
ijson==3.2.3
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
kiwisolver==1.4.5
label-studio==1.12.1
label-studio-converter==0.0.58
label-studio-tools==0.0.4
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
nltk==3.6.7
numpy==1.26.4
openai==1.30.5
opencv-python==4.9.0.80
ordered-set==4.0.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==5.27.1
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.15
pyparsing==3.1.2
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.1
redis==3.5.3
regex==2024.5.15
requests==2.31.0
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.1
scipy==1.13.1
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.3.1
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
ultralytics==8.2.28
ultralytics-thop==0.2.7
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
webencodings==0.5.1
wheel==0.40.0
xmljson==0.2.0
